Oraganized language processing e2e tests.
