+++ tests/cypress/integration/language-processing/classify-content-openapi-embeddings.test.js
@@ -1,0 +1,156 @@
+describe( '[Language processing] Classify Content (OpenAI) Tests', () => {
+	it( 'Can save OpenAI Embeddings "Language Processing" settings', () => {
+		cy.visit(
+			'/wp-admin/tools.php?page=classifai&tab=language_processing&provider=openai_embeddings'
+		);
+
+		cy.get( '#api_key' ).clear().type( 'password' );
+
+		cy.get( '#enable_classification' ).check();
+		cy.get( '#openai_embeddings_post_types_post' ).check();
+		cy.get( '#openai_embeddings_post_statuses_publish' ).check();
+		cy.get( '#openai_embeddings_taxonomies_category' ).check();
+		cy.get( '#number' ).clear().type( 1 );
+		cy.get( '#submit' ).click();
+	} );
+
+	it( 'Can create category and post and category will get auto-assigned', () => {
+		// Remove custom taxonomies so those don't interfere with the test.
+		cy.visit(
+			'/wp-admin/tools.php?page=classifai&tab=language_processing&provider=watson_nlu'
+		);
+		cy.get( '#classifai-settings-category' ).uncheck();
+		cy.get( '#classifai-settings-keyword' ).uncheck();
+		cy.get( '#classifai-settings-entity' ).uncheck();
+		cy.get( '#classifai-settings-concept' ).uncheck();
+		cy.get( '#submit' ).click();
+
+		// Create test term.
+		cy.deleteAllTerms( 'category' );
+		cy.createTerm( 'Test', 'category' );
+
+		// Create test post.
+		cy.createPost( {
+			title: 'Test embeddings',
+			content: 'Test embeddings content',
+		} );
+
+		// Close post publish panel.
+		const closePanelSelector = 'button[aria-label="Close panel"]';
+		cy.get( 'body' ).then( ( $body ) => {
+			if ( $body.find( closePanelSelector ).length > 0 ) {
+				cy.get( closePanelSelector ).click();
+			}
+		} );
+
+		// Open post settings sidebar.
+		cy.openDocumentSettingsSidebar();
+
+		// Find and open the category panel.
+		const panelButtonSelector = `.components-panel__body .components-panel__body-title button:contains("Categories")`;
+
+		cy.get( panelButtonSelector ).then( ( $panelButton ) => {
+			// Find the panel container.
+			const $panel = $panelButton.parents( '.components-panel__body' );
+
+			// Open panel.
+			if ( ! $panel.hasClass( 'is-opened' ) ) {
+				cy.wrap( $panelButton ).click();
+			}
+
+			// Ensure our test category is checked.
+			cy.wrap( $panel )
+				.find(
+					'.editor-post-taxonomies__hierarchical-terms-list .editor-post-taxonomies__hierarchical-terms-choice:first input'
+				)
+				.should( 'be.checked' );
+			cy.wrap( $panel )
+				.find(
+					'.editor-post-taxonomies__hierarchical-terms-list .editor-post-taxonomies__hierarchical-terms-choice:first label'
+				)
+				.contains( 'Test' );
+		} );
+	} );
+
+	it( 'Can create category and post and category will not get auto-assigned if feature turned off', () => {
+		cy.visit(
+			'/wp-admin/tools.php?page=classifai&tab=language_processing&provider=openai_embeddings'
+		);
+		cy.get( '#enable_classification' ).uncheck();
+		cy.get( '#submit' ).click();
+
+		// Create test term.
+		cy.deleteAllTerms( 'category' );
+		cy.createTerm( 'Test', 'category' );
+
+		// Create test post.
+		cy.createPost( {
+			title: 'Test embeddings disabled',
+			content: 'Test embeddings content',
+		} );
+
+		// Close post publish panel.
+		const closePanelSelector = 'button[aria-label="Close panel"]';
+		cy.get( 'body' ).then( ( $body ) => {
+			if ( $body.find( closePanelSelector ).length > 0 ) {
+				cy.get( closePanelSelector ).click();
+			}
+		} );
+
+		// Open post settings sidebar.
+		cy.openDocumentSettingsSidebar();
+
+		// Find and open the category panel.
+		const panelButtonSelector = `.components-panel__body .components-panel__body-title button:contains("Categories")`;
+
+		cy.get( panelButtonSelector ).then( ( $panelButton ) => {
+			// Find the panel container.
+			const $panel = $panelButton.parents( '.components-panel__body' );
+
+			// Open panel.
+			if ( ! $panel.hasClass( 'is-opened' ) ) {
+				cy.wrap( $panelButton ).click();
+			}
+
+			// Ensure our test category is not checked.
+			cy.wrap( $panel )
+				.find(
+					'.editor-post-taxonomies__hierarchical-terms-list .editor-post-taxonomies__hierarchical-terms-choice:first input'
+				)
+				.should( 'be.checked' );
+			cy.wrap( $panel )
+				.find(
+					'.editor-post-taxonomies__hierarchical-terms-list .editor-post-taxonomies__hierarchical-terms-choice:first label'
+				)
+				.contains( 'Uncategorized' );
+		} );
+	} );
+
+	it( 'Can see the enable button in a post (Classic Editor)', () => {
+		cy.visit( '/wp-admin/plugins.php' );
+		cy.get( '#activate-classic-editor' ).click();
+
+		cy.visit(
+			'/wp-admin/tools.php?page=classifai&tab=language_processing&provider=openai_embeddings'
+		);
+
+		cy.get( '#enable_classification' ).check();
+		cy.get( '#openai_embeddings_post_types_post' ).check();
+		cy.get( '#openai_embeddings_post_statuses_publish' ).check();
+		cy.get( '#openai_embeddings_taxonomies_category' ).check();
+		cy.get( '#number' ).clear().type( 1 );
+		cy.get( '#submit' ).click();
+
+		cy.classicCreatePost( {
+			title: 'Embeddings test classic',
+			content: "This feature uses OpenAI's Embeddings capabilities.",
+			postType: 'post',
+		} );
+
+		cy.get( '#classifai_language_processing_metabox' ).should( 'exist' );
+		cy.get( '#classifai-process-content' ).check();
+
+		cy.visit( '/wp-admin/plugins.php' );
+		cy.get( '#deactivate-classic-editor' ).click();
+	} );
+} );
